% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{runMediationBMDApp}
\alias{runMediationBMDApp}
\title{Launch the MediationBMDSearch Shiny App}
\usage{
runMediationBMDApp()
}
\description{
This function starts the Shiny app for searching mediation effects on BMD.
}
